+++ store/tikv/async_commit_fail_test.go
@@ -16,10 +16,12 @@ package tikv
 import (
 	"bytes"
 	"context"
+	"sort"
 
 	. "github.com/pingcap/check"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/failpoint"
+	"github.com/pingcap/kvproto/pkg/kvrpcpb"
 	"github.com/pingcap/parser/terror"
 	"github.com/pingcap/tidb/config"
 	"github.com/pingcap/tidb/kv"
@@ -80,4 +82,84 @@ func (s *testAsyncCommitFailSuite) TestFailAsyncCommitPrewriteRpcErrors(c *C) {
 	res, err := t2.Get(context.Background(), []byte("a"))
 	c.Assert(err, IsNil)
 	c.Assert(bytes.Equal(res, []byte("a1")), IsTrue)
+}
+
+func (s *testAsyncCommitSuite) TestSecondaryListInPrimaryLock(c *C) {
+	config.UpdateGlobal(func(conf *config.Config) {
+		conf.TiKVClient.EnableAsyncCommit = true
+	})
+	defer config.RestoreFunc()()
+
+	s.putAlphabets(c)
+
+	// Split into several regions.
+	for _, splitKey := range []string{"h", "o", "u"} {
+		loc, err := s.store.GetRegionCache().LocateKey(s.bo, []byte(splitKey))
+		c.Assert(err, IsNil)
+		newRegionID := s.cluster.AllocID()
+		newPeerID := s.cluster.AllocID()
+		s.cluster.Split(loc.Region.GetID(), newRegionID, []byte(splitKey), []uint64{newPeerID}, newPeerID)
+		s.store.GetRegionCache().InvalidateCachedRegion(loc.Region)
+	}
+
+	// Ensure the region has been split
+	loc, err := s.store.GetRegionCache().LocateKey(s.bo, []byte("i"))
+	c.Assert(err, IsNil)
+	c.Assert([]byte(loc.StartKey), BytesEquals, []byte("h"))
+	c.Assert([]byte(loc.EndKey), BytesEquals, []byte("o"))
+
+	loc, err = s.store.GetRegionCache().LocateKey(s.bo, []byte("p"))
+	c.Assert(err, IsNil)
+	c.Assert([]byte(loc.StartKey), BytesEquals, []byte("o"))
+	c.Assert([]byte(loc.EndKey), BytesEquals, []byte("u"))
+
+	var connID uint64 = 0
+	test := func(keys []string, values []string) {
+		connID++
+		ctx := context.WithValue(context.Background(), sessionctx.ConnID, connID)
+
+		txn, err := s.store.Begin()
+		c.Assert(err, IsNil)
+		for i := range keys {
+			txn.Set([]byte(keys[i]), []byte(values[i]))
+		}
+
+		c.Assert(failpoint.Enable("github.com/pingcap/tidb/store/tikv/asyncCommitDoNothing", "return"), IsNil)
+
+		err = txn.Commit(ctx)
+		c.Assert(err, IsNil)
+
+		tikvTxn := txn.(*tikvTxn)
+		primary := tikvTxn.committer.primary()
+		txnStatus, err := s.store.lockResolver.getTxnStatus(s.bo, txn.StartTS(), primary, 0, 0, false)
+		c.Assert(err, IsNil)
+		c.Assert(txnStatus.IsCommitted(), IsFalse)
+		c.Assert(txnStatus.action, Equals, kvrpcpb.Action_NoAction)
+		// Currently when the transaction has no secondary, the `secondaries` field of the txnStatus
+		// will be set nil. So here initialize the `expectedSecondaries` to nil too.
+		var expectedSecondaries [][]byte
+		for _, k := range keys {
+			if !bytes.Equal([]byte(k), primary) {
+				expectedSecondaries = append(expectedSecondaries, []byte(k))
+			}
+		}
+		sort.Slice(expectedSecondaries, func(i, j int) bool {
+			return bytes.Compare(expectedSecondaries[i], expectedSecondaries[j]) < 0
+		})
+
+		gotSecondaries := txnStatus.primaryLock.GetSecondaries()
+		sort.Slice(gotSecondaries, func(i, j int) bool {
+			return bytes.Compare(gotSecondaries[i], gotSecondaries[j]) < 0
+		})
+
+		c.Assert(gotSecondaries, DeepEquals, expectedSecondaries)
+
+		c.Assert(failpoint.Disable("github.com/pingcap/tidb/store/tikv/asyncCommitDoNothing"), IsNil)
+	}
+
+	test([]string{"a"}, []string{"a1"})
+	test([]string{"a", "b"}, []string{"a2", "b2"})
+	test([]string{"a", "b", "d"}, []string{"a3", "b3", "d3"})
+	test([]string{"a", "b", "h", "i", "u"}, []string{"a4", "b4", "h4", "i4", "u4"})
+	test([]string{"i", "a", "z", "u", "b"}, []string{"i5", "a5", "z5", "u5", "b5"})
 }
